Split subnet-specific startup into separate file
